+++ CHANGELOG.md
@@ -136,6 +136,7 @@
 - PR #3387 Strings column gather function
 - PR #3389 Move quantiles.hpp + group_quantiles.hpp files to legacy
 - PR #3398 Move reshape.hpp files to legacy
+- PR #3422 Move utilities to legacy
 
 ## Bug Fixes